Take clippy to zero across all three workspaces
`make lint` reached clippy for the first time in this branch and found
246 diagnostics. Most were mechanical and went through `clippy --fix`,
but three groups needed judgement rather than the suggested rewrite.

`--all-features` turns on `room-surface-profile`, which replaces the
normal cached-room submit path wholesale. That orphans nine helpers and
one parameter in psx-engine, so clippy called them dead and `--fix`
happily renamed a live parameter to `_shade_prewarmed_packets`, breaking
the default build. They are live in every shipping build, so they carry
`#[cfg_attr(feature = "room-surface-profile", ...)]` instead of being
deleted. `psx-gpu-render` had the mirror case: helpers used only from
`#[cfg(test)]` code, which the lib target cannot see, now gated on test.
`asset_streaming`'s `compact` is a third kind again -- its own doc says
it is deliberately uncalled, reserved for a level-change boundary where
no parsed view is live -- so it says that in an allow rather than
disappearing.

`--fix` also dropped an `#[inline(always)]` from psx-asset's hand-written
MIPS Q11 decoder. The function carried both `#[inline]` and
`#[inline(always)]` with the doc comment between them, and clippy removed
the second; the stray first one is gone instead, so the branchless
six-instruction form keeps the attribute it was written for.

The rest: `rust-version` was declared 1.75 while emulator-core already
used `repeat_n` (1.82) and `is_multiple_of` (1.87), and nothing builds
this tree on 1.75 anyway, so it now states the real floor. 71 field
reassignments became struct-update syntax, six complex tuple types got
named aliases, and the loop-index lints were split between genuine
rewrites and the cases where the counter indexes three to five parallel
arrays and an iterator buys nothing.

+++ engine/crates/psx-game-runtime/src/particles.rs
@@ -357,7 +357,7 @@ pub fn draw_room_atmosphere_overlay(
     let mut i = 0u32;
     while i < count {
         let seed = atmosphere_seed(i);
-        let layer = ((seed >> 4) & 3) as u32;
+        let layer = (seed >> 4) & 3;
         let fall_q4 = base_fall_q4 + (layer as i32) * 3;
         let wind_q4 = base_wind_q4 + layer as i32;
         let base_x = (seed & 0x1ff) as i32;
